resource type : inherit description, refactor code
